fix: make script modular
